refactor: move controller in utils
