+++ e-token/src/processor/utils/mod.rs
@@ -1,4 +1,6 @@
 mod asserts;
+mod group_receipt_controller;
 mod token;
 
+pub use group_receipt_controller::*;
 pub use token::*;
